Consolidated `auth` test utilities into main `tests/conftest.py` and removed redundant imports.
