+++ terraform/modules/lambdas/functions/tests/lambda_handlers/conftest.py
@@ -1,4 +1,3 @@
-from tests.utilities.auth.conftest import *
 from copy import deepcopy
 
 import pytest
